+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added more methods to `Expr`
 - Added more methods to `LazyFrame`
 - Added more methods to `DataFrame`
+- Added more methods to `Series`
 - Fixed `subset` option for `drop_nulls` method
 
 ## 0.21.0 (2025-08-03)